+++ bin/lips.js
@@ -1,17 +1,23 @@
 #!/usr/bin/env node
 
-const lily = require('@jcubic/lily');
-const util = require('util');
+import lily from '@jcubic/lily';
 
-const boolean = ['d', 'dynamic', 'q', 'quiet', 'V', 'version', 'trace', 't', 'debug', 'c', 'compile', 'b'];
+const boolean = ['d', 'dynamic', 'q', 'quiet', 'V', 'version', 'trace', 't', 'debug', 'c', 'compile', 'b', 'cbor'];
 const options = lily(process.argv.slice(2), { boolean });
+
+const quiet = options.q || options.quiet;
+
+import lips from '../src/lips.js';
 
 const {
     exec,
+    compile,
     parse,
     Formatter,
     serialize,
     unserialize,
+    serialize_bin,
+    unserialize_bin,
     balanced_parenthesis,
     tokenize,
     Interpreter,
@@ -26,16 +32,21 @@ const {
     env,
     banner,
     InputPort,
-    OutputPort } = require('../src/lips');
-const fs = require('fs');
-const os = require('os');
-const path = require('path');
-const { format } = require('util');
-const readline = require('readline');
-var highlight = require('prism-cli');
-var Prism = require('prismjs');
-require('prismjs/components/prism-scheme.min.js');
-require('../lib/js/prism.js');
+    OutputPort } = lips;
+
+import fs from 'fs';
+import os from 'os';
+import path from 'path';
+import { format } from 'util';
+import readline from 'readline';
+import highlight from 'prism-cli';
+import Prism from 'prismjs';
+import 'prismjs/components/prism-scheme.min.js';
+import '../lib/js/prism.js';
+
+import { createRequire } from 'module';
+
+const require = createRequire(import.meta.url);
 
 const kDebounceHistoryMS = 15;
 
@@ -107,8 +118,8 @@ function print(result) {
 
 // -----------------------------------------------------------------------------
 function bootstrap(interpreter) {
-    var list = ['./dist/std.scm'];
-    function open(name) {
+    const file = options.bootstrap ? options.bootstrap : './dist/std.xcb';
+    function read(name) {
         var path;
         try {
             path = require.resolve(`./${name}`);
@@ -119,25 +130,33 @@ function bootstrap(interpreter) {
                 path = require.resolve(`@jcubic/lips/../${name}`);
             }
         }
-        return fs.readFileSync(path).toString();
+        return readCode(path);
     }
-    return (function next() {
-        var name = list.shift();
-        if (name) {
-            var data = open(name);
-            if (compiled(name)) {
-                data = unserialize(data);
-            }
-            return run(data, interpreter, false, env.__parent__, true).then(next);
-        } else {
-            return Promise.resolve();
+    const code = read(file);
+    return run(code, interpreter, false, env.__parent__, true);
+}
+
+// -----------------------------------------------------------------------------
+function readCode(filename) {
+    if (compiled_binary(filename)) {
+        return unserialize_bin(readBinary(filename));
+    } else {
+        const code = readFile(filename);
+        if (compiled(filename)) {
+            return unserialize(code);
         }
-    })();
+        return code;
+    }
 }
 
 // -----------------------------------------------------------------------------
 function compiled(name) {
     return name.match(/\.xcm$/);
+}
+
+// -----------------------------------------------------------------------------
+function compiled_binary(name) {
+    return name.match(/\.xcb$/);
 }
 
 // -----------------------------------------------------------------------------
@@ -177,6 +196,9 @@ function log(message) {
 var strace;
 var rl;
 var newline;
+const moduleURL = new URL(import.meta.url);
+const __dirname = path.dirname(moduleURL.pathname);
+const __filename = path.basename(moduleURL.pathname);
 var interp = Interpreter('repl', {
     stdin: InputPort(function() {
         return new Promise(function(resolve) {
@@ -238,12 +260,17 @@ var interp = Interpreter('repl', {
     '__help': env.get('help')
 });
 
+function readBinary(filename) {
+    return fs.readFileSync(filename);
+}
+
+// -----------------------------------------------------------------------------
 function readFile(filename) {
-    const buff = fs.readFileSync(filename);
+    const buff = readBinary(filename);
     return buff.toString().replace(/^#!.*\n/, '');
 }
-// -----------------------------------------------------------------------------
 
+// -----------------------------------------------------------------------------
 if (options.version || options.V) {
     // SRFI 176
     global.output = Pair.fromArray([
@@ -272,23 +299,39 @@ if (options.version || options.V) {
     });
 } else if (options.b && options._.length === 1) {
     const filename = options._[0];
-    const code = readFile(filename);
-    if (compiled(filename)) {
-        for (expr of unserialize(code)) {
-            console.log(expr.toString(true));
-        }
+    for (expr of readCode(filename)) {
+        console.log(expr.toString(true));
     }
 } else if ((options.c || options.compile) && options._.length === 1) {
     try {
         const filename = options._[0];
-        console.log('Experimental compiler');
-        console.log(`Compiling ${filename} ...`);
-        const compiled_name = filename.replace(/\.[^.]+$/, '') + '.xcm';
+        if (!quiet) {
+            console.log('Experimental compiler');
+            console.log(`Compiling ${filename} ...`);
+        }
+        const ext = '.xcb';
+        const compiled_name = filename.replace(/\.[^.]+$/, '') + ext;
         var code = readFile(filename);
         bootstrap(interp).then(function() {
-            parse(code, interp.__env__).then(code => {
-                fs.writeFileSync(compiled_name, serialize(code));
-            });
+            return compile(code, interp.__env__).then(code => {
+                if (!quiet) {
+                    console.log(`Writing ${compiled_name} ...`);
+                }
+                try {
+                    const encoded = serialize_bin(code);
+                    fs.writeFile(compiled_name, encoded, function(err) {
+                        if (err) {
+                            console.error(err);
+                        } else if (!quiet) {
+                            console.log('DONE');
+                        }
+                    });
+                } catch (e) {
+                    console.error(e);
+                }
+            })
+        }).catch(e => {
+            print_error(e, true);
         });
     } catch(e) {
         console.log(e);
@@ -310,11 +353,8 @@ if (options.version || options.V) {
     });
     const filename = options._[0];
     try {
-        let code = readFile(filename);
-        if (compiled(filename)) {
-            code = unserialize(code);
-        }
-        return bootstrap(interp).then(() => {
+        const code = readCode(filename);
+        bootstrap(interp).then(() => {
             const dynamic = options.d || options.dynamic;
             return run(code, interp, dynamic, null, options.t || options.trace);
         });
@@ -337,7 +377,7 @@ if (options.version || options.V) {
 } else {
     const dynamic = options.d || options.dynamic;
     const entry = '   ' + (dynamic ? 'dynamic' : 'lexical') + ' scope $1';
-    if (process.stdin.isTTY && !options.q && !options.quiet) {
+    if (process.stdin.isTTY && !quiet) {
         console.log(banner.replace(/(\n\nLIPS.+)/m, entry)); // '
     }
     var prompt = 'lips> ';
@@ -380,11 +420,11 @@ function run_repl(err, rl) {
     var multiline = false;
     var resolve;
     // we use promise loop to fix issue when copy paste list of S-Expression
-    var prev_eval = Promise.resolve();
+    let prev_eval = Promise.resolve();
     if (process.stdin.isTTY) {
         rl.prompt();
     }
-    var prev_line;
+    let prev_line;
     bootstrap(interp).then(function() {
         rl.on('line', function(line) {
             code += line;
@@ -412,7 +452,7 @@ function run_repl(err, rl) {
                     }
                     return '\x1b[K' + prefix + line;
                 }).join('\n');
-                let num = lines.length + count;
+                let num = lines.length + count + 1;
                 const format = `\x1b[${num}F${stdout}\n`;
                 process.stdout.write(format);
             }
@@ -425,7 +465,7 @@ function run_repl(err, rl) {
                     rl.setPrompt('');
                     rl.pause();
                     prev_eval = prev_eval.then(function() {
-                        var result = run(code, interp, dynamic, null, options.t || options.trace);
+                        const result = run(code, interp, dynamic, null, options.t || options.trace);
                         code = '';
                         return result;
                     }).then(function(result) {
@@ -456,10 +496,10 @@ function run_repl(err, rl) {
                     });
                 } else {
                     multiline = true;
-                    var ind = indent(code, 2, prompt.length - continuePrompt.length);
+                    const ind = indent(code, 2, prompt.length - continuePrompt.length);
                     rl.setPrompt(continuePrompt);
                     rl.prompt();
-                    spaces = new Array(ind + 1).join(' ');
+                    const spaces = new Array(ind + 1).join(' ');
                     if (terminal) {
                         rl.write(spaces);
                     }