+++ rollup.config.js
@@ -1,16 +1,18 @@
-import commonjs from "rollup-plugin-commonjs";
-import nodeResolve from "rollup-plugin-node-resolve";
-import babel from "rollup-plugin-babel";
+import commonjs from "@rollup/plugin-commonjs";
+import nodeResolve from "@rollup/plugin-node-resolve";
+import babel from "@rollup/plugin-babel";
+import nodePolyfills from 'rollup-plugin-node-polyfills';
 
 export default {
     input: "src/lips.js",
     output: {
-        name: "lib",
+        name: "lips",
         file: "dist/lips.js",
-        format: "iife",
+        format: "umd",
         globals: {
             "@babel/runtime/regenerator": "regeneratorRuntime"
-        }
+        },
+        manualChunks: () => 'everything.js'
     },
     onwarn: (warning, next) => {
         const str = warning.toString();
@@ -22,7 +24,7 @@ export default {
     plugins: [
         babel({
             "babelrc": false,
-            "runtimeHelpers": true,
+            babelHelpers: 'runtime',
             "plugins": [
                 "@babel/plugin-transform-async-to-generator",
                 "@babel/plugin-transform-regenerator",
@@ -36,11 +38,12 @@ export default {
             ],
             "exclude": "node_modules/**"
         }),
-        nodeResolve({
-            mainFields: ["jsnext:main"]
-        }),
+        nodePolyfills(),
         commonjs({
             include: "node_modules/**"
+        }),
+        nodeResolve({
+            mainFields: ["jsnext:main"]
         })
     ]
 };